[refactor] Isolate classloader in separate package as it is externally used too
