+++ exist-core/src/main/java/org/exist/test/ExistEmbeddedServer.java
@@ -24,8 +24,8 @@ package org.exist.test;
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
 import org.exist.EXistException;
-import org.exist.start.Classpath;
-import org.exist.start.EXistClassLoader;
+import org.exist.start.classloader.Classpath;
+import org.exist.start.classloader.EXistClassLoader;
 import org.exist.storage.BrokerPool;
 import org.exist.storage.journal.Journal;
 import org.exist.util.Configuration;
