+++ exist-start/src/main/java/org/exist/start/classloader/Classpath.java
@@ -36,7 +36,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.exist.start;
+package org.exist.start.classloader;
 
 import java.io.File;
 import java.net.MalformedURLException;
