+++ exist-start/src/main/java/org/exist/start/classloader/EXistClassLoader.java
@@ -19,7 +19,7 @@
  * License along with this library; if not, write to the Free Software
  * Foundation, Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02110-1301  USA
  */
-package org.exist.start;
+package org.exist.start.classloader;
 
 import java.net.MalformedURLException;
 import java.net.URL;
